add repository responses

+++ ToDoList.Application/Services/ToDoListService.cs
@@ -1,4 +1,5 @@
-﻿using ToDoList.Application.Interfaces.Repositories;
+﻿using Microsoft.Extensions.Logging;
+using ToDoList.Application.Interfaces.Repositories;
 using ToDoList.Application.Interfaces.Services;
 using ToDoList.Domain.Entities;
 
@@ -7,10 +8,12 @@ namespace ToDoList.Application.Services
     public class ToDoListService : IToDoListService
     {
         private readonly IToDoListRepository _toDoListRepository;
+        private readonly ILogger<ToDoListService> _logger;
 
-        public ToDoListService(IToDoListRepository toDoListRepository)
+        public ToDoListService(IToDoListRepository toDoListRepository, ILogger<ToDoListService> logger)
         {
             _toDoListRepository = toDoListRepository;
+            _logger = logger;
         }
 
         public Task<IReadOnlyCollection<ToDoItem>> GetToDoItemsAsync(CancellationToken cancellation)
@@ -18,19 +21,37 @@ namespace ToDoList.Application.Services
             return _toDoListRepository.GetToDoItemsAsync(cancellation);
         }
 
-        public Task CreateToDoItemAsync(ToDoItem item, CancellationToken cancellation)
+        public async Task<bool> CreateToDoItemAsync(ToDoItem item, CancellationToken cancellation)
         {
-            return _toDoListRepository.CreateToDoItemAsync(item, cancellation);
+            if (item == null)
+            {
+                _logger.LogWarning("The Item being created must not be null!");
+                return false;
+            }
+
+            return await _toDoListRepository.CreateToDoItemAsync(item, cancellation);
         }
 
-        public Task DeleteToDoItemAsync(int id, CancellationToken cancellation)
+        public async Task<bool> DeleteToDoItemAsync(int id, CancellationToken cancellation)
         {
-            return _toDoListRepository.DeleteToDoItemAsync(id, cancellation);
+            if (id == 0)
+            {
+                _logger.LogWarning("The Item id being deleted must not be 0!");
+                return false;
+            }
+
+            return await _toDoListRepository.DeleteToDoItemAsync(id, cancellation);
         }
 
-        public Task UpdateToDoItemAsync(ToDoItem item, CancellationToken cancellation)
+        public async Task<bool> UpdateToDoItemAsync(ToDoItem item, CancellationToken cancellation)
         {
-            return _toDoListRepository.UpdateToDoItemAsync(item, cancellation);
+            if (item == null)
+            {
+                _logger.LogWarning("The Item being updated must not be null!");
+                return false;
+            }
+
+            return await _toDoListRepository.UpdateToDoItemAsync(item, cancellation);
         }
     }
 }
